+++ ui/app/components/clients/page/overview.hbs
@@ -30,5 +30,5 @@
       {{/each}}
     </S.Options>
   </Hds::Form::Select::Base>
-  <Clients::Table @data={{this.tableData}} @filteredMonth={{this.selectedMonth}} />
+  <Clients::Table @data={{this.tableData}} />
 {{/if}}
