+++ ui/app/components/clients/page/overview.ts
@@ -5,17 +5,13 @@
 
 import ActivityComponent from '../activity';
 import { service } from '@ember/service';
-import type FlagsService from 'vault/services/flags';
 import { tracked } from '@glimmer/tracking';
 import { action } from '@ember/object';
 import { HTMLElementEvent } from 'vault/forms';
 import { parseAPITimestamp } from 'core/utils/date-formatters';
-import { MountClients } from 'core/utils/client-count-utils';
-import RouterService from '@ember/routing/router-service';
-
-interface TableData extends MountClients {
-  namespace: string;
-}
+import { formatTableData, TableData } from 'core/utils/client-count-utils';
+import type FlagsService from 'vault/services/flags';
+import type RouterService from '@ember/routing/router-service';
 
 export default class ClientsOverviewPageComponent extends ActivityComponent {
   @service declare readonly flags: FlagsService;
@@ -36,23 +32,9 @@ export default class ClientsOverviewPageComponent extends ActivityComponent {
     }));
   }
 
-  get tableData(): TableData[] {
-    if (!this.selectedMonth) return [];
-    // get data from selected month
-    const monthData = this.byMonthNewClients.find((m) => m.month === this.selectedMonth);
-    const namespaces = monthData?.namespaces;
-
-    let data: TableData[] = [];
-    // iterate over namespaces to add "namespace" to each mount object
-    namespaces?.forEach((n) => {
-      const mounts: TableData[] = n.mounts.map((m) => {
-        // add namespace to mount block
-        return { ...m, namespace: n.label };
-      });
-      data = [...data, ...mounts];
-    });
-
-    return data;
+  get tableData(): TableData[] | undefined {
+    if (!this.selectedMonth) return undefined;
+    return formatTableData(this.byMonthNewClients, this.selectedMonth);
   }
 
   @action
