+++ ui/app/templates/vault/cluster/clients/counts/overview.hbs
@@ -10,5 +10,4 @@
   @endTimestamp={{this.model.endTimestamp}}
   @namespace={{this.countsController.ns}}
   @mountPath={{this.countsController.mountPath}}
-  @updateQueryParams={{this.countsController.updateQueryParams}}
 />
